# specs/001-push-notification-latency/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Push Notification Latency Coordinator API
  version: "1.0.0"
  description: >
    Generic, platform-agnostic coordinator API for enrolling devices, triggering
    measurement rounds, recording receipts, and reporting latency/success-rate
    statistics. See specs/001-push-notification-latency/spec.md.
servers:
  - url: /v1
paths:
...
  /reports/success-rate:
    get:
      summary: Delivery success rate, sliceable by round send time (FR-014)
      operationId: successRateReport
      parameters:
        - name: bucket
          in: query
          description: Time bucket to group rounds by
          schema:
            type: string
            enum: [hour, day]
            default: hour
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success-rate buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuccessRateBucket'
components:
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
      description: Static operator/scheduler bearer token (FR-029); stored as a secret, not in source (FR-030).
  schemas:
...
    SuccessRateBucket:
      type: object
      properties:
        bucket_start:
          type: string
          format: date-time
        targeted:
          type: integer
        received:
          type: integer
        success_rate:
          type: number
          format: float